programOpenList and program open 2 components
